refactor: converter and keyword tests split
